Added scene objects to level 1 and finalized collision with artifacts
Added some simple trees to the first level and starting area. Also, the rotation issue on top of the elevator is now fixed.

I also fixed the collision with artifacts. So when a player picks up an artifact, the collision is now handled by using CapsuleComponent instead of BoxCollider. That way you don't have to kick the artifact or walk with the right foot first towards the artifact.