Added altar and artifact dropoff
Added the altar and functionality of dropping off the artifacts to the altar. The altar also holds the artifacts now.

Also added an interact button for the player. So when it is close to any object that is assign to the Interacting collision channel, the player can interact with the object