Fixed some bugs for AI attack and started balancing the attack of AI and player
The AI had some serious bugs related to their attack. All are now fixed. I also started on creating a more smooth attack for the player and AI. Right now I have only implemented it on the SpiritEnemy, but I will do the same for the player and Shaman later