Added special ability for spirit enemy
The spiritenemy can now "chase" the player while he/she is in the physical world, slow him/her down and call the decrement lives function from the character.

I'm still struggling a bit with making the character move back to its home location, but it shouldn't be too hard. It's most likely a problem in the Behaviour Tree but not entirely sure how to fix it yet