Fixed two bugs related to the spirit enemy
The game crashed if you tried to kick the head after the enemy dies. Fixed that by switching the heads collision fra Query&Physics to PhysicsOnly.

The other bug was that the enemy stopped following after the player sometimes, if the playerkicked the spirit enemy at the exact right time. It made the task enemyAttack jump out of the task if the enum RunningAnimation is equal to TakingDamage