Added floating head trailer
Made the head float next to the player and switch between the mask and head depending on what world you are in.

I couldn't get the timeline to work for some reason. The SwitchingHeadTimelineFloatReturn seem to have some bad coding. Don't know what is happening but the heads move very weirdly around