Set up the cut scene (NB!!!!!! READ DESCRIPTION)
Remember to download the .mp4 file Aztec Puppet.mp4 from the drive. Rename it to IntroCustScene.mp4 before you open the project. Then build the project again. If this isn't done, the custscene won't work.

The reason why this was done is because the .mp4 file is too large for githuband it is too late for us to get started on LFS